requirements.txt
pytelegrambotapi
mysql-connector